+++ translations/caesiumph_en_US.ts
@@ -1,0 +1,463 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="en_US">
+<context>
+    <name>AboutDialog</name>
+    <message>
+        <location filename="../aboutdialog.ui" line="29"/>
+        <location filename="../aboutdialog.ui" line="121"/>
+        <source>CaesiumPH</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../aboutdialog.ui" line="136"/>
+        <source>Lossless JPEG compression</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../aboutdialog.ui" line="200"/>
+        <source>© 2016 - SaeraSoft and Matteo Paonessa</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../aboutdialog.ui" line="215"/>
+        <source>All rights reserved.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../aboutdialog.ui" line="230"/>
+        <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;a href=&quot;http://saerasoft.com/caesium&quot;&gt;&lt;span style=&quot; text-decoration: underline; color:#0000ff;&quot;&gt;License&lt;/span&gt;&lt;/a&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../aboutdialog.cpp" line="52"/>
+        <source>Version</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
+    <name>CaesiumPH</name>
+    <message>
+        <location filename="../caesiumph.ui" line="14"/>
+        <location filename="../caesiumph.cpp" line="532"/>
+        <location filename="../caesiumph.cpp" line="675"/>
+        <source>CaesiumPH</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="106"/>
+        <source>Name</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="117"/>
+        <source>Original Size</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="122"/>
+        <source>New Size</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="127"/>
+        <source>Saved</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="132"/>
+        <source>Full Path</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="230"/>
+        <source>Add files</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="276"/>
+        <source>Import folder</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="328"/>
+        <location filename="../caesiumph.ui" line="734"/>
+        <source>Compress</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="377"/>
+        <location filename="../caesiumph.cpp" line="783"/>
+        <source>Remove item</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="426"/>
+        <location filename="../caesiumph.ui" line="724"/>
+        <location filename="../caesiumph.cpp" line="798"/>
+        <source>Clear list</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="501"/>
+        <source>File</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="514"/>
+        <source>Edit</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="523"/>
+        <source>Actions</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="529"/>
+        <source>Help</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="615"/>
+        <source>&lt;!DOCTYPE HTML PUBLIC &quot;-//W3C//DTD HTML 4.0//EN&quot; &quot;http://www.w3.org/TR/REC-html40/strict.dtd&quot;&gt;
+&lt;html&gt;&lt;head&gt;&lt;meta name=&quot;qrichtext&quot; content=&quot;1&quot; /&gt;&lt;style type=&quot;text/css&quot;&gt;
+p, li { white-space: pre-wrap; }
+&lt;/style&gt;&lt;/head&gt;&lt;body style=&quot; font-family:&apos;.SF NS Text&apos;; font-size:13pt; font-weight:400; font-style:normal;&quot;&gt;
+&lt;p style=&quot; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;span style=&quot; font-family:&apos;.Helvetica Neue DeskInterface&apos;;&quot;&gt;No EXIF info available&lt;/span&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="695"/>
+        <source>Add pictures...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="700"/>
+        <source>Add folder...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="705"/>
+        <source>Quit</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="716"/>
+        <source>Remove items</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="729"/>
+        <source>Select all</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="739"/>
+        <source>About CaesiumPH...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="750"/>
+        <source>Save list</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="758"/>
+        <location filename="../caesiumph.cpp" line="840"/>
+        <source>Save list as...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="761"/>
+        <source>Save list as</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.ui" line="766"/>
+        <source>Open list...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="113"/>
+        <source>No EXIF info available</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="130"/>
+        <source>A new version is available!</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="292"/>
+        <location filename="../caesiumph.cpp" line="869"/>
+        <source>Import files...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="304"/>
+        <source>Importing...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="304"/>
+        <source>Cancel</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="361"/>
+        <location filename="../caesiumph.cpp" line="363"/>
+        <source> files added to the list</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="362"/>
+        <source> duplicates found</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="369"/>
+        <source>Select a folder to import...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="391"/>
+        <source> items removed</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="533"/>
+        <source>Compressing...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="575"/>
+        <source>Compression completed! </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="576"/>
+        <source> files compressed in </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="578"/>
+        <source>from </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="578"/>
+        <source> to </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="579"/>
+        <source>Saved </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="676"/>
+        <source>Do you really want to exit?</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="729"/>
+        <source>Update available</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="730"/>
+        <source>This will close CaesiumPH. Do you really want to update now?</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="752"/>
+        <source> files in list</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="756"/>
+        <source>List cleared</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="784"/>
+        <source>Remove the item from the list</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="788"/>
+        <source>Show in folder</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="789"/>
+        <source>Opens the folder containing the file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="793"/>
+        <source>Show destination folder</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="794"/>
+        <source>Opens the destination folder for the file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../caesiumph.cpp" line="799"/>
+        <source>Clears the list</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
+    <name>PreferenceDialog</name>
+    <message>
+        <location filename="../preferencedialog.ui" line="17"/>
+        <source>Preferences</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="135"/>
+        <source>Language</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="155"/>
+        <source>Overwrite original files</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="168"/>
+        <source>Leaving empty will overwrite the file</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="96"/>
+        <source>Custom suffix</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="101"/>
+        <source>Subfolder</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="106"/>
+        <source>Custom folder</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="115"/>
+        <source>-- System language --</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="195"/>
+        <source>Prompt before exit</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="208"/>
+        <source>Walk through subfolders</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="223"/>
+        <source>Language will change after restart</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="256"/>
+        <source>Keep copyright</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="279"/>
+        <source>Keep date</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="315"/>
+        <source>Progressive</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="325"/>
+        <source>Keep comments</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="332"/>
+        <source>Keep EXIF information</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="397"/>
+        <source>Send anonymous usage information</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="413"/>
+        <source>See info...</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="425"/>
+        <source>Not yet implemented</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="477"/>
+        <source>General</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="486"/>
+        <source>Compression</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.ui" line="495"/>
+        <source>Privacy</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.cpp" line="131"/>
+        <source>This data will help improve this application and won&apos;t be shared with anyone.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.cpp" line="147"/>
+        <source>_compressed</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.cpp" line="150"/>
+        <source>compressed</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../preferencedialog.cpp" line="164"/>
+        <source>Select an output folder...</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
+    <name>QIODevice</name>
+    <message>
+        <location filename="../caesiumph.h" line="102"/>
+        <source>Image Files</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
+    <name>Utils</name>
+    <message>
+        <location filename="../utils.cpp" line="158"/>
+        <source> minutes</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+</TS>